+++ src/core/Recoil_Batching.js
@@ -10,7 +10,9 @@
  */
 
 const {batchStart} = require('../core/Recoil_RecoilValueInterface');
-const {unstable_batchedUpdates} = require('../util/Recoil_ReactBatchedUpdates');
+const {
+  unstable_batchedUpdates,
+} = require('../util/polyfill/ReactBatchedUpdates');
 
 let batcher = unstable_batchedUpdates;
