+++ src/util/polyfill/ReactBatchedUpdates.js
